tests/pytest_docker_compose_tests/my_network/a_buildable_container/requirements.txt
uvicorn==0.27.1
psycopg2-binary==2.9.9
fastapi==0.110.0
